+++ wp-express-checkout/includes/integrations/simple-wp-membership/index.php
@@ -1,0 +1,1 @@
+<?php // Silence is golden
