+++ wp-express-checkout/tests/includes/integrations/emember/test-class-emember.php
@@ -1,0 +1,82 @@
+<?php
+
+namespace WP_Express_Checkout\Integrations;
+
+use WP_Express_Checkout\Products;
+use WP_UnitTestCase;
+
+/**
+ * Generated by PHPUnit_SkeletonGenerator on 2021-06-28 at 07:52:47.
+ *
+ * @group integrations
+ *
+ * @covers WP_Express_Checkout\Integrations\Emember
+ */
+class EmemberTest extends WP_UnitTestCase {
+
+	/**
+	 * @var Emember
+	 */
+	protected $object;
+
+	/**
+	 * Sets up the fixture, for example, opens a network connection.
+	 * This method is called before a test is executed.
+	 */
+	public function setUp() {
+		require_once WPEC_TESTS_DIR . '/mocks/mock-emember-functions.php';
+		require_once WPEC_TESTS_DIR . '/mocks/mock-emember-auth.php';
+		$this->object = new Emember;
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Integrations\Emember::handle_signup
+	 */
+	public function testHandle_emember_signup() {
+		$payment = json_decode( file_get_contents( WPEC_TESTS_DIR . '/data/payment-data.json' ), true );
+		$product = $this->factory->post->create(
+			[
+				'post_type' => Products::$products_slug,
+				'meta_input' => [
+					'wpec_product_emember_level' => 42,
+				]
+			]
+		);
+
+		$this->object->handle_signup( $payment, 0, $product );
+
+		$this->expectOutputString( 'eMember_handle_subsc_signup_stand_alone' );
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Integrations\Emember::handle_signup
+	 */
+	public function testHandle_emember_signup__no_level() {
+		$this->object->handle_signup( [], 0, 0 );
+		$this->expectOutputString( '' );
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Integrations\Emember::get_member_info_from_api
+	 */
+	public function testGet_member_info_from_api() {
+		$payment = json_decode( file_get_contents( WPEC_TESTS_DIR . '/data/payment-data.json' ), true );
+		$ipn_data = $this->object->get_member_info_from_api( $payment );
+
+		$this->assertEqualSetsWithIndex(
+			$ipn_data,
+			[
+				'payer_email'     => 'customer@example.com',
+				'first_name'      => 'John',
+				'last_name'       => 'Doe',
+				'txn_id'          => '5O190127TN364715T',
+				'address_street'  => '2211 N First Street',
+				'address_city'    => 'San Jose',
+				'address_state'   => 'CA',
+				'address_zip'     => '95131',
+				'address_country' => 'United States',
+			]
+		);
+	}
+
+}
