+++ wp-express-checkout/tests/includes/integrations/simple-wp-membership/test-class-simple-wp-membership.php
@@ -1,0 +1,62 @@
+<?php
+
+namespace WP_Express_Checkout\Integrations;
+
+use WP_Express_Checkout\Products;
+use WP_UnitTestCase;
+
+/**
+ * Generated by PHPUnit_SkeletonGenerator on 2021-06-28 at 07:52:47.
+ *
+ * @group integrations
+ *
+ * @covers WP_Express_Checkout\Integrations\Simple_WP_Membership
+ */
+class Simple_WP_MembershipTest extends WP_UnitTestCase {
+
+	/**
+	 * @var Simple_WP_Membership
+	 */
+	protected $object;
+
+	/**
+	 * Sets up the fixture, for example, opens a network connection.
+	 * This method is called before a test is executed.
+	 */
+	public function setUp() {
+		require_once WPEC_TESTS_DIR . '/mocks/mock-swmp-functions.php';
+		require_once WPEC_TESTS_DIR . '/mocks/mock-swpm-utils.php';
+		require_once WPEC_TESTS_DIR . '/mocks/mock-swpm-utils-member.php';
+
+		$this->object = new Simple_WP_Membership();
+	}
+
+
+	/**
+	 * @covers WP_Express_Checkout\Integrations\Simple_WP_Membership::handle_signup
+	 */
+	public function testHandle_swpm_signup() {
+		$payment = json_decode( file_get_contents( WPEC_TESTS_DIR . '/data/payment-data.json' ), true );
+		$product = $this->factory->post->create(
+			[
+				'post_type' => Products::$products_slug,
+				'meta_input' => [
+					'wpec_product_swpm_level' => 42,
+				]
+			]
+		);
+
+		$this->object->handle_signup( $payment, 0, $product );
+
+		$this->expectOutputString( 'swpm_handle_subsc_signup_stand_alone' );
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Integrations\Simple_WP_Membership::handle_signup
+	 */
+	public function testHandle_swpm_signup__no_level() {
+		$this->object->handle_signup( [], 0, 0 );
+		$this->expectOutputString( '' );
+	}
+
+}
